+++ EnumExtencions/EnumExtencions.sln
@@ -1,20 +1,20 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31613.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnumExtencions", "EnumExtencions\EnumExtencions.csproj", "{80DF6CC3-D757-4A86-B40F-A91050BED761}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumExtencions", "EnumExtencions\EnumExtencions.csproj", "{80DF6CC3-D757-4A86-B40F-A91050BED761}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Release|Any CPU.Build.0 = Release|Any CPU
+		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Debug|x64.ActiveCfg = Debug|x64
+		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Debug|x64.Build.0 = Debug|x64
+		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Release|x64.ActiveCfg = Release|x64
+		{80DF6CC3-D757-4A86-B40F-A91050BED761}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
