+++ EnumExtencions/EnumExtencions/EnumExtencions.csproj
@@ -3,6 +3,11 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
+    <Platforms>x64</Platforms>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
+    <DocumentationFile>\EnumExtencions\EnumExtencions\bin\DocumentationFile.xml</DocumentationFile>
   </PropertyGroup>
 
 </Project>
